requirements.txt
chainlit==1.0.502
langchain==0.1.14
openai==1.16.2
python-dotenv==1.0.1
